All Done I Guess v2.0

+++ Admin/courses/courseDelete.php
@@ -1,6 +1,12 @@
 <?php
 include '../../connect.php';
 
+ 
+if (!isset($_SESSION['isAdmin']) || $_SESSION['isAdmin'] != 1) {
+    header('Location: ../../Home.php');
+    exit();
+}
+
 if(isset($_GET['ID'])) {
   $id = $_GET['ID'];
   $sql = "DELETE FROM courses WHERE ID = $id";
