+++ Admin/timetable/timetableDelete.php
@@ -1,6 +1,11 @@
 <?php
 include '../../connect.php';
 
+if (!isset($_SESSION['isAdmin']) || $_SESSION['isAdmin'] != 1) {
+    header('Location: ../../Home.php');
+    exit();
+}
+
 if(isset($_GET['ID'])) {
   $id = $_GET['ID'];
   $sql = "DELETE FROM timetable WHERE ID = $id";
